requirements-dev.txt
# Development and Testing Requirements for SSH Profile Manager

# Testing Framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-xdist>=3.3.1
pytest-benchmark>=4.0.0

# Code Quality and Linting
flake8>=6.0.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0

# Flake8 Extensions
flake8-docstrings>=1.7.0
flake8-bugbear>=23.9.16
flake8-comprehensions>=3.14.0
flake8-simplify>=0.20.0

# Coverage Reporting
coverage[toml]>=7.3.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Security
bandit>=1.7.5
safety>=2.3.0

# Mock and Testing Utilities
freezegun>=1.2.2
parameterized>=0.9.0

# CLI Testing
click-testing>=0.1.1

# Build and Release Tools
build>=1.0.0
twine>=4.0.0
wheel>=0.41.0

# Pre-commit Hooks
pre-commit>=3.5.0

# Type Stubs
types-all>=1.0.0

# Development Tools
tox>=4.11.0
pip-tools>=7.3.0

# Additional Testing Utilities
responses>=0.23.0
httpx>=0.25.0
